+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Encrpyted SQL*Net connection if server has SQLNET.ENCRYPTION_SERVER != rejected<br>
   Requires release of v6.1.7 for https://github.com/rsim/oracle-enhanced as precondition (https://github.com/rsim/oracle-enhanced/pull/2284)
 - Alternative Ruby client library for Kafka removed, using only Apache Kafka client library now
+- The number of created change events at initialization is recorded now in activity logs
 
 ## 2025-04-03 1.11.1
 - The default library controlled by KAFKA_CLIENT_LIBRARY is changed to 'java' in the next release. The ruby-kafka library will be removed in a future release.